+++ src/rp2_common/hardware_claim/claim.c
@@ -20,7 +20,7 @@ bool hw_is_claimed(uint8_t *bits, uint bit_index) {
     if (bits[bit_index >> 3u] & (1u << (bit_index & 7u))) {
         rc = false;
     } else {
-        bits[bit_index >> 3u] |= (1u << (bit_index & 7u));
+        bits[bit_index >> 3u] |= (uint8_t)(1u << (bit_index & 7u));
         rc = true;
     }
     hw_claim_unlock(save);
@@ -32,7 +32,7 @@ void hw_claim_or_assert(uint8_t *bits, uint bit_index, const char *message) {
     if (bits[bit_index >> 3u] & (1u << (bit_index & 7u))) {
         panic(message, bit_index);
     } else {
-        bits[bit_index >> 3u] |= (1u << (bit_index & 7u));
+        bits[bit_index >> 3u] |= (uint8_t)(1u << (bit_index & 7u));
     }
     hw_claim_unlock(save);
 }
@@ -43,8 +43,8 @@ int hw_claim_unused_from_range(uint8_t *bits, bool required, uint bit_lsb, uint 
     int found_bit = -1;
     for(uint bit=bit_lsb; bit <= bit_msb; bit++) {
         if (!(bits[bit >> 3u] & (1u << (bit & 7u)))) {
-            bits[bit >> 3u] |= (1u << (bit & 7u));
-            found_bit = bit;
+            bits[bit >> 3u] |= (uint8_t)(1u << (bit & 7u));
+            found_bit = (int)bit;
             break;
         }
     }
@@ -58,7 +58,7 @@ int hw_claim_unused_from_range(uint8_t *bits, bool required, uint bit_lsb, uint 
 void hw_claim_clear(uint8_t *bits, uint bit_index) {
     uint32_t save = hw_claim_lock();
     assert(bits[bit_index >> 3u] & (1u << (bit_index & 7u)));
-    bits[bit_index >> 3u] &= ~(1u << (bit_index & 7u));
+    bits[bit_index >> 3u] &= (uint8_t) ~(1u << (bit_index & 7u));
     hw_claim_unlock(save);
 }
